refactor(ruby/mpp): mirror Rust spine layout
Reshapes ruby/lib/mpp/ to match rust/crates/mpp/src/, addressing Ludo's
"similar for the mpp module" review on PR #127:

  ruby/lib/mpp/
    error.rb, expires.rb, store.rb, version.rb     stays (top-level utility)
    challenge.rb, settlement.rb                     stays (public 402/200 wrappers)
    sinatra.rb                                      stays (framework shim)
    protocol/
      core/
        challenge.rb           was mpp/core/challenge.rb
        credential.rb          was mpp/core/credential.rb
        receipt.rb             was mpp/core/receipt.rb
        headers.rb             was mpp/headers.rb (top-level)
        challenge_store.rb     was mpp/core/challenge_store.rb (was mpp/internal/)
      intents/
        charge.rb              was mpp/intent/charge_request.rb
      solana.rb                was mpp/methods/solana.rb
      solana/
        verifier.rb            was mpp/methods/solana/verifier.rb
        verification_result.rb was mpp/methods/solana/verification_result.rb
    server/
      charge.rb                merged mpp/server.rb (Server::Instance) and
                               mpp/core/handler.rb. Exposes Mpp::Server::Charge
                               (was Server::Instance) with nested
                               Mpp::Server::Charge::Handler (was Mpp::Core::Handler).
      middleware.rb            stays
      decorator.rb             stays

Module renames (no backward-compat shims):

  Mpp::Methods::Solana          -> Mpp::Protocol::Solana
  Mpp::Headers                  -> Mpp::Protocol::Core::Headers
  Mpp::Intent::ChargeRequest    -> Mpp::Protocol::Intents::ChargeRequest
  Mpp::Core::Challenge          -> Mpp::Protocol::Core::Challenge
  Mpp::Core::Credential         -> Mpp::Protocol::Core::Credential
  Mpp::Core::Receipt            -> Mpp::Protocol::Core::Receipt
  Mpp::Core::ChallengeStore     -> Mpp::Protocol::Core::ChallengeStore
  Mpp::Core::Handler            -> Mpp::Server::Charge::Handler
  Mpp::Server::Instance         -> Mpp::Server::Charge

Mpp.create unchanged (public factory). 259 tests, 0 failures.
