+++ ruby/lib/pay_core/headers.rb
@@ -3,7 +3,7 @@
 module PayCore
   # Generic HTTP auth-scheme + auth-param parser per RFC 7235 sec 2.1
   # and 4.1, shared by solana-mpp and solana-x402. Protocol-specific
-  # bindings (e.g. constructing `Mpp::Core::Challenge` from a parsed
+  # bindings (e.g. constructing `Mpp::Protocol::Core::Challenge` from a parsed
   # `Payment` challenge) live in their respective layers; this module
   # only owns the tokenisation, quote-aware splitting, escaping, and
   # auth-param key/value parsing.
